RUN-3241 moved IAB tests out of the multi runtime file
